requirements.txt
# Core
numpy>=1.26
pandas>=2.2
polars>=0.20
scipy>=1.13

# Graph / Deep Learning
torch>=2.3
torch_geometric>=2.5
networkx>=3.3

# Bayesian
pymc>=5.14
arviz>=0.18

# MLOps
wandb>=0.17

# Scraping / HTTP
requests>=2.32
httpx>=0.27
playwright>=1.44
beautifulsoup4>=4.12
lxml>=5.2

# Data sources
sportsipy>=0.6
kaggle>=1.6
python-dotenv>=1.0

# Scheduling / caching
APScheduler>=3.10
diskcache>=5.6

# Testing
pytest>=8.2
pytest-cov>=5.0
pytest-mock>=3.14
responses>=0.25       # mock HTTP in tests

# API server
fastapi>=0.115
uvicorn[standard]>=0.30

# Utilities
tqdm>=4.66
loguru>=0.7
